+++ CHANGELOG.md
@@ -7,6 +7,9 @@ _The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Added
+- Generallize mandatory words and add mandatory_bio_keywords
+
 ### Changed 
 - Update xpath for like/unlike and comment